requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
MouseInfo==0.1.3
openai==1.12.0
pillow==10.2.0
PyAutoGUI==0.9.54
pydantic==2.6.1
pydantic-core==2.16.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
pytweening==1.1.0
rubicon-objc==0.4.7
sniffio==1.3.0
tqdm==4.66.2
typing-extensions==4.9.0
easyocr